+++ package.json
@@ -4,50 +4,39 @@
   "private": true,
   "homepage": "https://avdotion.github.io/clue",
   "dependencies": {
-    "@babel/core": "7.7.4",
-    "@svgr/webpack": "4.3.3",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.4.0",
-    "@testing-library/user-event": "^7.2.1",
-    "@types/jest": "^24.0.25",
-    "@types/node": "^12.12.24",
-    "@types/react": "^16.9.17",
-    "@types/react-dom": "^16.9.4",
-    "@types/react-redux": "^7.1.5",
-    "@typescript-eslint/eslint-plugin": "^2.8.0",
-    "@typescript-eslint/parser": "^2.8.0",
+    "@babel/core": "7.7.7",
+    "@svgr/webpack": "5.0.1",
+    "@typescript-eslint/eslint-plugin": "^2.15.0",
+    "@typescript-eslint/parser": "^2.15.0",
     "babel-eslint": "10.0.3",
     "babel-jest": "^24.9.0",
     "babel-loader": "8.0.6",
     "babel-plugin-named-asset-import": "^0.3.5",
     "babel-preset-react-app": "^9.1.0",
-    "camelcase": "^5.3.1",
+    "camelcase": "^5.2.0",
     "case-sensitive-paths-webpack-plugin": "2.2.0",
     "crypto-es": "^1.1.1",
-    "css-loader": "3.2.0",
+    "css-loader": "3.4.1",
     "dotenv": "8.2.0",
     "dotenv-expand": "5.1.0",
-    "eslint": "^6.6.0",
+    "eslint": "^6.8.0",
     "eslint-config-react-app": "^5.1.0",
-    "eslint-loader": "3.0.2",
-    "eslint-plugin-flowtype": "3.13.0",
-    "eslint-plugin-import": "2.18.2",
+    "eslint-loader": "3.0.3",
+    "eslint-plugin-flowtype": "4.5.3",
+    "eslint-plugin-import": "2.19.1",
     "eslint-plugin-jsx-a11y": "6.2.3",
-    "eslint-plugin-react": "7.16.0",
-    "eslint-plugin-react-hooks": "^1.6.1",
-    "file-loader": "4.3.0",
-    "fs-extra": "^8.1.0",
-    "gh-pages": "^2.1.1",
+    "eslint-plugin-react": "7.17.0",
+    "eslint-plugin-react-hooks": "^2.3.0",
+    "file-loader": "5.0.2",
+    "fs-extra": "8.1.0",
     "html-webpack-plugin": "4.0.0-beta.5",
-    "husky": "^3.1.0",
     "identity-obj-proxy": "3.0.0",
-    "immer": "^5.1.0",
+    "is-wsl": "2.1.1",
     "jest": "24.9.0",
-    "jest-environment-jsdom-fourteen": "0.1.0",
+    "jest-environment-jsdom-fourteen": "1.0.1",
     "jest-resolve": "24.9.0",
     "jest-watch-typeahead": "0.4.2",
-    "lint-staged": "^9.5.0",
-    "mini-css-extract-plugin": "0.8.0",
+    "mini-css-extract-plugin": "0.9.0",
     "normalize.css": "^8.0.1",
     "optimize-css-assets-webpack-plugin": "5.0.3",
     "pnp-webpack-plugin": "1.5.0",
@@ -58,25 +47,31 @@
     "postcss-safe-parser": "4.0.1",
     "react": "^16.12.0",
     "react-app-polyfill": "^1.0.5",
-    "react-dev-utils": "^10.0.0",
+    "react-dev-utils": "10.0.0",
     "react-dom": "^16.12.0",
-    "react-redux": "^7.1.3",
-    "redux": "^4.0.5",
     "reshadow": "0.0.1-alpha.73",
-    "resolve": "1.12.2",
+    "resolve": "1.14.1",
     "resolve-url-loader": "3.1.1",
     "sass-loader": "8.0.0",
-    "semver": "6.3.0",
-    "style-loader": "1.0.0",
-    "terser-webpack-plugin": "2.2.1",
+    "semver": "7.1.1",
+    "style-loader": "1.1.2",
+    "terser-webpack-plugin": "2.3.1",
     "ts-pnp": "1.1.5",
-    "typesafe-actions": "^5.1.0",
-    "typescript": "^3.7.4",
-    "url-loader": "2.3.0",
-    "webpack": "4.41.2",
-    "webpack-dev-server": "3.9.0",
+    "url-loader": "3.0.0",
+    "webpack": "4.41.5",
+    "webpack-dev-server": "3.10.1",
     "webpack-manifest-plugin": "2.2.0",
     "workbox-webpack-plugin": "4.3.1"
+  },
+  "devDependencies": {
+    "typescript": "3.7.4",
+    "@types/jest": "24.0.25",
+    "@types/node": "12.11.7",
+    "@types/react": "16.9.17",
+    "@types/react-dom": "16.9.4",
+    "husky": "^3.0.9",
+    "lint-staged": "^9.4.2",
+    "gh-pages": "^2.1.1"
   },
   "scripts": {
     "start": "node scripts/start.js",
@@ -151,9 +146,7 @@
     "setupFiles": [
       "react-app-polyfill/jsdom"
     ],
-    "setupFilesAfterEnv": [
-      "<rootDir>/src/setupTests.ts"
-    ],
+    "setupFilesAfterEnv": [],
     "testMatch": [
       "<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}",
       "<rootDir>/src/**/*.{spec,test}.{js,jsx,ts,tsx}"
@@ -193,15 +186,6 @@
   "babel": {
     "presets": [
       "react-app"
-    ],
-    "plugins": [
-      [
-        "reshadow/babel",
-        {
-          "postcss": "true",
-          "files": "/.shadow.css$/"
-        }
-      ]
     ]
   }
 }
