+++ src/entities/secretData.ts
@@ -1,0 +1,5 @@
+export type SecretData = {
+  masterPassword: string,
+  domainName: string,
+  optionalSalt: string,
+};
