--- src/store/rootAction.ts
@@ -1,13 +1,0 @@
-import * as AutoCopyActions from '#/widgets/AutoCopy/actions';
-import * as DomainNameActions from '#/widgets/DomainName/actions';
-import * as HashMethodActions from '#/widgets/HashMethod/actions';
-import * as MasterPasswordActions from '#/widgets/MasterPassword/actions';
-import * as OptionalSaltActions from '#/widgets/OptionalSalt/actions';
-
-export default {
-  AutoCopy: AutoCopyActions,
-  DomainName: DomainNameActions,
-  HashMethod: HashMethodActions,
-  MasterPassword: MasterPasswordActions,
-  OptionalSalt: OptionalSaltActions,
-};
