--- src/widgets/AutoCopy/types.ts
@@ -1,5 +1,0 @@
-export type AutoCopy = boolean;
-
-export type State = {
-  isEnabled: AutoCopy,
-};
