--- src/widgets/DomainName/actions.ts
@@ -1,8 +1,0 @@
-import {createAction} from 'typesafe-actions';
-
-import {DomainName} from './types';
-
-export const setValue = createAction(
-  '@DomainName/SET_VALUE',
-  (newDomainName: DomainName) => newDomainName
-)<DomainName>();
