--- src/widgets/DomainName/index.ts
@@ -1,19 +1,0 @@
-import {connect} from 'react-redux';
-
-import {RootState} from '#/store';
-
-import View from './view';
-import {setValue} from './actions';
-
-const mapStateToProps = ({domainName: {value}}: RootState) => ({
-  domainName: value,
-});
-
-const mapDispatchToProps = {
-  setDomainName: setValue,
-};
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(View);
