--- src/widgets/DomainName/types.ts
@@ -1,5 +1,0 @@
-export type DomainName = string;
-
-export type State = {
-  value: DomainName,
-};
