--- src/widgets/HashMethod/types.ts
@@ -1,7 +1,0 @@
-import {HashMethod} from '#/utils/crypto';
-
-export type State = {
-  currentMethod: HashMethod,
-  methods: readonly HashMethod[],
-};
-
