--- src/widgets/MasterPassword/actions.ts
@@ -1,8 +1,0 @@
-import {createAction} from 'typesafe-actions';
-
-import {MasterPassword} from './types';
-
-export const setValue = createAction(
-  '@MasterPassword/SET_VALUE',
-  (newMasterPassword: MasterPassword) => newMasterPassword
-)<MasterPassword>();
