--- src/widgets/MasterPassword/index.ts
@@ -1,19 +1,0 @@
-import {connect} from 'react-redux';
-
-import {RootState} from '#/store';
-
-import View from './view';
-import {setValue} from './actions';
-
-const mapStateToProps = ({masterPassword: {value}}: RootState) => ({
-  masterPassword: value,
-});
-
-const mapDispatchToProps = {
-  setMasterPassword: setValue,
-};
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(View);
