--- src/widgets/MasterPassword/types.ts
@@ -1,5 +1,0 @@
-export type MasterPassword = string;
-
-export type State = {
-  value: MasterPassword,
-};
