--- src/widgets/OptionalSalt/actions.ts
@@ -1,8 +1,0 @@
-import {createAction} from 'typesafe-actions';
-
-import {OptionalSalt} from './types';
-
-export const setValue = createAction(
-  '@OptionalSalt/SET_VALUE',
-  (newOptionalSalt: OptionalSalt) => newOptionalSalt
-)<OptionalSalt>();
