--- src/widgets/OptionalSalt/types.ts
@@ -1,5 +1,0 @@
-export type OptionalSalt = string;
-
-export type State = {
-  value: OptionalSalt,
-};
