+++ src/widgets/protection-bar.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
-import styled, {use} from 'reshadow';
-
-import {MasterPassword} from '#/widgets/MasterPassword/types';
-import {DomainName} from '#/widgets/DomainName/types';
-import {OptionalSalt} from '#/widgets/OptionalSalt/types';
+// @ts-ignore
+import styled, {use} from 'reshadow/macro';
+import {SecretData} from './../entities/secretData';
 
 // eslint-disable-next-line max-len
 const EMOJI_LIST = ['😺','😸','😹','😻','😼','😽','🙀','😿','😾','🙈','🙉','🦝','🐵','🐒','🦍','🐶','🐕','🐩','🐺','🦊','🐱','🐈','🦁','🐯','🐅','🐆','🐴','🐎','🦄','🦓','🦌','🐮','🐂','🦙','🐃','🐄','🐷','🦛','🐖','🐗','🐽','🐏','🐑','🐐','🐪','🐫','🦒','🐘','🦏','🐭','🐁','🐀','🦘','🐹','🦡','🐰','🐇','🐿','🦔','🦇','🐻','🐨','🐼','🐾','🦃','🐔','🦢','🐓','🐣','🐤','🦚','🐥','🐦','🦜','🐧','🕊','🦅','🦆','🦉','🐸','🐊','🐢','🦎','🐍','🐲','🐉','🦕','🦖','🐳','🐋','🐬','🐟','🐠','🐡','🦈','🐙','🐚','🦀','🦟','🦐','🦠','🦑','🐌','🦋','🐛','🐜','🐝','🐞','🦗','🕷','🕸', '🦂','🦞','🌸','💮','🏵','🌹','🥀','🌺','🌻','🌼','🌷','🌱','🌲','🌳','🌴','🌵','🌾','🌿','☘','🍀','🍁','🍂','🍃','🍄','💅','👓','🕶','👔','👕','👖','🧣','🧤','🧥','🧦','👗','👘','👙','👚','👛','👜','👝','🛍','🎒','👞','👟','👠','👡','👢','👑','👒','🎩','🎓','🧢','⛑','📿','💄','🌂','☂','🎽','🥽','🥼','🥾','🥿','🧺','🚂','🚃','🚄','🚅','🚆','🚇','🚈','🚉','🚊','🚝','🚞','🚋','🚌','🚍','🚎','🚐','🚑','🚒','🚓','🚔','🚕','🚖','🚗','🚘','🚙','🚚','🚛','🚜','🚲','🛴','🛵','🚏','🛣','🛤','⛵','🛶','🚤','🛳','⛴','🛥','🚢','✈','🛩','🛫','🛬','🚁','🚟','🚠','🚡','🛰','🚀','🛸','🌍','🌎','🌏','🌐','🗺','🗾','🏔','⛰','🗻','🏕','🏖','🏜','🏝','🏞','🏟','🏛','🏗','🏘','🏚','🏠','🏡','🏢','🏣','🏤','🏥','🏦','🏨','🏩','🏪','🏫','🏬','🏭','🏯','🏰','🗼','🗽','⛪','🕌','🕍','⛩','🕋','⛲','⛺','🏙','🎠','🎡','🎢','🎪','⛳','🗿','💦','🌋','🌁','🌃','🌄','🌅','🌆','🌇','🌉','🌌','🌑','🌒','🌓','🌔','🌕','🌖','🌗','🌘','🌙','🌚','🌛','🌜','🌡','☀','🌝','🌞','🌟','🌠','⛅','⛈','🌤','🌥','🌦','🌧','🌨','🌩','🌪','🌫','🌬','🌀','🌈','⛄','💧','🌊','🎑'];
@@ -17,39 +15,38 @@ const emoji = (value: string): string => {
 };
 
 type Props = {
-  masterPassword: MasterPassword,
-  domainName: DomainName,
-  optionalSalt: OptionalSalt,
+  secretData: SecretData,
 };
 
-export default function EmojiBarComponent ({
-  masterPassword,
-  domainName,
-  optionalSalt,
-}: Props) {
-  return styled`
-    |wrapper {
-      line-height: 2.6em;
-      font-size: 3.4em;
-      letter-spacing: 0.35em;
-      text-align: center;
-      width: 100%;
-    }
-  `(
-    <use.wrapper>{
-      masterPassword === ''
-      ? new Array(3).fill(DEFAULT_EMOJI)
-      : domainName === ''
-        ? [
+export const DataVisualizationBar: React.FC<Props> = ({
+  secretData: {
+    masterPassword,
+    domainName,
+    optionalSalt,
+  },
+}: Props) => styled`
+  |wrapper {
+    line-height: 2.6em;
+    font-size: 3.4em;
+    letter-spacing: 0.35em;
+    text-align: center;
+    width: 100%;
+  }
+`(
+  <use.wrapper>{
+    masterPassword === ''
+    ? new Array(3).fill(DEFAULT_EMOJI)
+    : domainName === ''
+      ? [
+        emoji(masterPassword),
+        DEFAULT_EMOJI,
+        DEFAULT_EMOJI,
+      ]
+      : [
           emoji(masterPassword),
-          DEFAULT_EMOJI,
-          DEFAULT_EMOJI,
+          emoji(domainName + optionalSalt),
+          emoji(optionalSalt + domainName + 'shift'),
         ]
-        : [
-            emoji(masterPassword),
-            emoji(domainName + optionalSalt),
-            emoji(optionalSalt + domainName + 'shift'),
-          ]
-    }</use.wrapper>
-  );
-}
+  }</use.wrapper>
+);
+
