+++ tsconfig.json
@@ -17,13 +17,7 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "jsx": "react",
-    "baseUrl": "src",
-    "paths": {
-      "#/*": [
-        "./*"
-      ]
-    }
+    "jsx": "react"
   },
   "include": [
     "src"
